requirements.txt
openai
faiss-cpu
pandas
numpy
sentence-transformers
spacy
streamlit
